pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mixmatch"
version = "0.2.0"
description = "Extract BPM, sections, and keys from local audio files using signal processing. No external APIs required."
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Weston Carpenter", email = "wc.cc96@gmail.com"}
]
keywords = [
    "music",
    "audio-analysis",
    "bpm",
    "key-detection",
    "signal-processing",
    "librosa"
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Sound/Audio",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    "librosa>=0.10.0",
    "numpy>=1.24.0",
    "soundfile>=0.12.0",
    "click>=8.1.0",
    "openai-whisper>=20231117",
    "demucs>=4.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
    "isort>=5.12.0",
]
docs = [
    "sphinx>=6.0.0",
    "sphinx-rtd-theme>=1.2.0",
]

[project.scripts]
mixmatch = "mixmatch.cli:main"

[tool.setuptools]
packages = ["mixmatch"]

[tool.black]
line-length = 100
target-version = ['py310']

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=mixmatch --cov-report=html --cov-report=term-missing"
